OnlineEditor機能を、コンポーネントに切り出す
